+++ docs/src/release_notes.rst
@@ -18,6 +18,7 @@ Unreleased
 - Resolve string annotations in the module scope (:issue:`62`)
 - Correctly ensure that return annotations are valid types (:issue:`63`)
 - Resolve imported members to their original location (:issue:`64`)
+- Fix multi-target assignment and unpacked assignment (:issue:`66`)
 
 0.5.1 (2021-11-20)
 ------------------